Changed file name and added conditional rendering for responsiveness
